refactor(solver): split tuple display alias helpers
